+++ code/midtown/mmaudio/crevbuf.h
@@ -57,7 +57,7 @@ public:
 
 private:
     // 0x4F4B10 | ?QueryProperty@CReverbBuffer@@AAEHK@Z
-    ARTS_IMPORT i32 QueryProperty(u32 arg1);
+    ARTS_IMPORT i32 QueryProperty(ulong arg1);
 
     u8 gap0[0x4];
 };
