+++ code/midtown/mmwidget/manager.cpp
@@ -33,7 +33,7 @@ void MenuManager::AddWidgets(class Bank* /*arg1*/)
 {}
 
 // 0x4B1B30 | ?HelpWatcher@@YGKPAX@Z
-ARTS_IMPORT /*static*/ u32 ARTS_STDCALL HelpWatcher(void* arg1);
+ARTS_IMPORT /*static*/ ulong ARTS_STDCALL HelpWatcher(void* arg1);
 
 i32 MenuManager::GetPreviousMenu()
 {
